public/Front_Kang/requirements.txt
fastapi==0.116.1
uvicorn==0.23.2
langchain==0.3.26
langchain-openai==0.3.28
redis==6.2.0
pydantic==2.11.7
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
